Skyler Sebold - Finished IMU dashboard + Swerve Module Dashboard
Finished and correctly converted IMU angle to a double for smart dashboard in the drive subsystem. Also added smart dashboard widgets for the SwerveModule Subsystem, although unknown if that is needed or if its would be better to have widgets on each module in the drive subsystem, but I had trouble figuring out how to properly implement that. Question still remains of if there must be code in the initialize, however I am pretty sure what is implemented now is satisfactory.