Skyler Sebold - Continuous Pivot PIDS
Removed IsFinished check from Pivot Command so it runs .SetReference() continuously for the PIDS in the execute section until interrupted where it then sets the motors to zero in the end section of the command. This allows the Pivot Arm to counteract the constant pull of gravity while it is needed to stay at its specified angle and position.